split test_deploy_app into 2 files
